pulled out cholesky, lower/upper triangular solve
