+++ sympy/matrices/tests/test_matrices.py
@@ -2866,7 +2866,7 @@ def test_upper_triangular_solve():
 
     raises(NonSquareMatrixError,
         lambda: Matrix([1, 0]).upper_triangular_solve(Matrix([0, 1])))
-    raises(TypeError,
+    raises(ShapeError,
         lambda: Matrix([[1, 0], [0, 1]]).upper_triangular_solve(Matrix([1])))
     raises(TypeError,
         lambda: Matrix([[2, 1], [1, 2]]).upper_triangular_solve(
